refactor(sound): improve sound function parameter handling and clean up unused variables
- Updated the S_OpenAL_StopLoopingSound function to handle unused parameters more explicitly.
- Removed the numOpenALSounds variable from snd_openal.c as it was no longer needed.
- Adjusted compatibility checks in backwards_compatibility.c to cast content_path correctly.
- Enhanced type safety in various functions by explicitly casting sizes to int.
- Cleaned up unused variable warnings in compatibility_shims.c and fallback_assets.c.

+++ src/common/net_protocol_validation.c
@@ -457,7 +457,7 @@ Validate entity state data
 ===============
 */
 net_protocol_result_t Net_ValidateEntityState(const byte *data, int length) {
-    if (!data || length < sizeof(entityState_t)) {
+    if (!data || length < (int)sizeof(entityState_t)) {
         return NET_PROTOCOL_INVALID_LENGTH;
     }
 
@@ -486,7 +486,7 @@ Validate player state data
 ===============
 */
 net_protocol_result_t Net_ValidatePlayerState(const byte *data, int length) {
-    if (!data || length < sizeof(playerState_t)) {
+    if (!data || length < (int)sizeof(playerState_t)) {
         return NET_PROTOCOL_INVALID_LENGTH;
     }
